Remove need to deserialize / serialize local messages
Typically, messages received from the local ROS master are automatically
deserialized.  However, to send to an external client they are then
serialized again.  By not deserializing messages when they are received,
we can avoid this overhead.  We do this by dynamically creating new class
types which replace the standard deserialization method with one that just
copies the bytestream to an internal variable.