Implement chunking of large UDP packets
If an outgoing message is larger than 1400 bytes (i.e. the MTU) the sender
splits it up into several smaller bytes and transmits each of those. The
receiver gathers these up and once it has received all of them, transmits
the full message.

There's also a simple retransmit protocol. If the receiver only gets some
of the packets, it sends a request to the sender containing a list of
packets that it's missing. The sender then retransmits these. The sender
saves chunked messages in a local cache which it periodically
purges. Similarly, the receiver will only request a retransmit for a
message for a certain amount of time.