Add success / failure info to subscription requests
Previously, when trying to subscribe to a message there was no way to know
if the callback was successful. Now, Topic.srv has an additional success
field which is true if and only if the remote end acknowledged the
request. The service request will fail if the remote end does not respond
within 5 seconds. Additionally, the local end will only create a publisher
for a foreign topic once its request has been acknowledged.

To support these changes, the UDP server no longer filters messages based
on the sequence number. All received messages will have the callback
executed, so messages that require acknowledgement should be idempotent.